requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
Django==5.2.7
django-ninja==1.4.5
orjson==3.11.4
pillow==12.0.0
pydantic==2.12.3
pydantic_core==2.41.4
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
